Machine learning project for Book recommendation System
This project has three section one is dataset which we use for our model.
Second is ML model building in which I have used Cosine similarity feature to develope the suyetm to generate the similar book.
Third and last file contains the guide to deployment of this project using html, CSS, Django and other libraries which u may have to install.